project_observer_package/pyproject.toml
[tool.poetry]
name = "project_observer"
version = "1.0.0"
description = "The User Interface for the FLOps plugin of Oakestra"
authors = ["Alexander Malyuk <malyuk.alexander1999@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "2.0"
icecream = "^2.1.3"
flops-utils = "^0.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
oak = "project_observer.main:main"

[tool.black]
line-length = 100
